rendezés

+++ src/app/pages/product-list/product-list.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
+import { Observable} from 'rxjs';
 
 import { Product } from 'src/app/model/product';
 import { ProductService } from 'src/app/service/product.service';
-import { Observable} from 'rxjs';
 
 @Component({
   selector: 'app-product-list',
@@ -11,15 +11,28 @@ import { Observable} from 'rxjs';
 })
 export class ProductListComponent implements OnInit {
 
-  productList$: Observable<Product[]> = this.productService.productList$;
+  productList$: Observable<Product[]> = this.productService.getAll();
+  phrase: string = '';
+  columnKey: string = '';
 
-  constructor(private productService: ProductService) { }
+  constructor(private productService: ProductService,) { }
 
   ngOnInit(): void {
     this.productService.getAll();
   }
 
+  onColumnSelect(key: string): void {
+    this.columnKey = key;    
+  }
+
+  onChangePhrase(event: any): void {
+    this.phrase = (event.target as HTMLInputElement).value;
+  }
+
   deleteItem(product: Product): void {
+    if (!window.confirm("Biztosan törölni akarod?")) {
+      return;
+    }
     this.productService.remove(product).subscribe(
       () => location.reload()
     );
